Reorganize backend files to make it futurproof
