+++ .github/workflows/pr.yml
@@ -35,7 +35,6 @@ jobs:
           context: ./backend
           push: true
           tags: ghcr.io/datalab-mi/basegun/basegun-backend:${{ github.head_ref }}
-          target: dev
 
   build-frontend:
     name: Build Frontend
@@ -55,7 +54,6 @@ jobs:
           context: ./frontend
           push: true
           tags: ghcr.io/datalab-mi/basegun/basegun-frontend:${{ github.head_ref }}
-          target: prod
   
   test-backend:
     name: Test Backend
