+++ backend/requirements.txt
@@ -8,10 +8,9 @@ gelf-formatter==0.2.1
 pyyaml>=5.4.1
 user-agents==2.2.0
 boto3==1.28.39
-torch==2.1.1
-torchvision==0.16.1
-ultralytics==8.1.2
 autodynatrace==2.0.0
+# ML
+ultralytics==8.1.2
 # Dev
 pytest==7.4.3
 coverage==7.3.2
