+++ backend/tests/test_api.py
@@ -7,7 +7,9 @@ import boto3
 import pytest
 import requests
 from fastapi.testclient import TestClient
-from src.main import S3_BUCKET_NAME, S3_URL_ENDPOINT, app
+
+from src.config import S3_BUCKET_NAME, S3_URL_ENDPOINT
+from src.main import app
 
 client = TestClient(app)
