+++ docker-compose.yml
@@ -9,7 +9,6 @@ services:
         - VERSION=${TAG:-latest}
         - CACERT_LOCATION
       context: ./backend
-      target: ${BUILD_TARGET:-dev}
     command: uvicorn src.main:app --reload --host 0.0.0.0 --port 5000 --no-server-header
     container_name: basegun-backend
     environment:
